pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "werkzeug>=3.1.3",
]
